Move session management helpers to tools/pytest_fixtures/webdriver.py
Issue: #59157
